migrate to hardhat, remove useless files and add test_bridge script

+++ contracts/DeusBridge.sol
@@ -30,61 +30,61 @@ contract DeusBridge is IDeusBridge, Ownable, Pausable {
 
     /* ========== STATE VARIABLES ========== */
 
-    uint public lastTxId = 0;  // unique id for deposit tx
-    uint public network;  // current chain id
-    uint public minReqSigs;  // minimum required tss
-    uint public scale = 1e6;
-    uint public bridgeReserve;  // it handles buyback & recollaterlize on dei pools
-    address public muonContract;  // muon signature verifier contract
+    uint256 public lastTxId = 0; // unique id for deposit tx
+    uint256 public network; // current chain id
+    uint256 public minReqSigs; // minimum required tss
+    uint256 public scale = 1e6;
+    uint256 public bridgeReserve; // it handles buyback & recollaterlize on dei pools
+    address public muonContract; // muon signature verifier contract
     address public deiAddress;
-    uint8   public ETH_APP_ID;  // muon's eth app id
-    bool    public mintable;  // use mint functions instead of transfer
+    uint8 public ETH_APP_ID; // muon's eth app id
+    bool public mintable; // use mint functions instead of transfer
     // we assign a unique ID to each chain (default is CHAIN-ID)
-    mapping (uint => address) public sideContracts;
+    mapping(uint256 => address) public sideContracts;
     // tokenId => tokenContractAddress
-    mapping(uint => address)  public tokens;
-    mapping(uint => Transaction) private txs;
+    mapping(uint256 => address) public tokens;
+    mapping(uint256 => Transaction) private txs;
     // user => (destination chain => user's txs id)
-    mapping(address => mapping(uint => uint[])) private userTxs;
+    mapping(address => mapping(uint256 => uint256[])) private userTxs;
     // source chain => (tx id => false/true)
-    mapping(uint => mapping(uint => bool)) public claimedTxs;
+    mapping(uint256 => mapping(uint256 => bool)) public claimedTxs;
     // tokenId => tokenFee
-    mapping(uint => uint) public fee;
+    mapping(uint256 => uint256) public fee;
     // tokenId => collectedFee
-    mapping(uint => uint) public collectedFee;
+    mapping(uint256 => uint256) public collectedFee;
 
     /* ========== EVENTS ========== */
     event Deposit(
         address indexed user,
-        uint tokenId,
-        uint amount,
-        uint indexed toChain,
-        uint txId
+        uint256 tokenId,
+        uint256 amount,
+        uint256 indexed toChain,
+        uint256 txId
     );
     event DepositWithReferralCode(
         address indexed user,
-        uint tokenId,
-        uint amount,
-        uint indexed toChain,
-        uint txId,
-        uint referralCode
+        uint256 tokenId,
+        uint256 amount,
+        uint256 indexed toChain,
+        uint256 txId,
+        uint256 referralCode
     );
     event Claim(
         address indexed user,
-        uint tokenId, 
-        uint amount, 
-        uint indexed fromChain, 
-        uint txId
+        uint256 tokenId,
+        uint256 amount,
+        uint256 indexed fromChain,
+        uint256 txId
     );
-    event BridgeReserveSet(uint bridgeReserve, uint _bridgeReserve);
+    event BridgeReserveSet(uint256 bridgeReserve, uint256 _bridgeReserve);
 
     /* ========== CONSTRUCTOR ========== */
 
     constructor(
-        uint minReqSigs_, 
-        uint bridgeReserve_,
+        uint256 minReqSigs_,
+        uint256 bridgeReserve_,
         uint8 ETH_APP_ID_,
-        address muon_, 
+        address muon_,
         address deiAddress_,
         bool mintable_
     ) {
@@ -100,56 +100,69 @@ contract DeusBridge is IDeusBridge, Ownable, Pausable {
     /* ========== PUBLIC FUNCTIONS ========== */
 
     function deposit(
-        uint amount, 
-        uint toChain,
-        uint tokenId
-    ) external returns (uint txId) {
+        uint256 amount,
+        uint256 toChain,
+        uint256 tokenId
+    ) external returns (uint256 txId) {
         txId = _deposit(msg.sender, amount, toChain, tokenId);
         emit Deposit(msg.sender, tokenId, amount, toChain, txId);
     }
 
     function depositFor(
         address user,
-        uint amount, 
-        uint toChain,
-        uint tokenId
-    ) external returns (uint txId) {
+        uint256 amount,
+        uint256 toChain,
+        uint256 tokenId
+    ) external returns (uint256 txId) {
         txId = _deposit(user, amount, toChain, tokenId);
         emit Deposit(user, tokenId, amount, toChain, txId);
     }
 
     function deposit(
-        uint amount, 
-        uint toChain,
-        uint tokenId,
-        uint referralCode
-    ) external returns (uint txId) {
+        uint256 amount,
+        uint256 toChain,
+        uint256 tokenId,
+        uint256 referralCode
+    ) external returns (uint256 txId) {
         txId = _deposit(msg.sender, amount, toChain, tokenId);
-        emit DepositWithReferralCode(msg.sender, tokenId, amount, toChain, txId, referralCode);
+        emit DepositWithReferralCode(
+            msg.sender,
+            tokenId,
+            amount,
+            toChain,
+            txId,
+            referralCode
+        );
     }
 
     function depositFor(
         address user,
-        uint amount, 
-        uint toChain,
-        uint tokenId,
-        uint referralCode
-    ) external returns (uint txId) {
+        uint256 amount,
+        uint256 toChain,
+        uint256 tokenId,
+        uint256 referralCode
+    ) external returns (uint256 txId) {
         txId = _deposit(user, amount, toChain, tokenId);
-        emit DepositWithReferralCode(user, tokenId, amount, toChain, txId, referralCode);
+        emit DepositWithReferralCode(
+            user,
+            tokenId,
+            amount,
+            toChain,
+            txId,
+            referralCode
+        );
     }
 
     function _deposit(
         address user,
-        uint amount,
-        uint toChain,
-        uint tokenId
-    ) 
-        internal 
-        whenNotPaused() 
-        returns (uint txId) 
-    {
-        require(sideContracts[toChain] != address(0), "Bridge: unknown toChain");
+        uint256 amount,
+        uint256 toChain,
+        uint256 tokenId
+    ) internal whenNotPaused returns (uint256 txId) {
+        require(
+            sideContracts[toChain] != address(0),
+            "Bridge: unknown toChain"
+        );
         require(toChain != network, "Bridge: selfDeposit");
         require(tokens[tokenId] != address(0), "Bridge: unknown tokenId");
 
@@ -164,7 +177,7 @@ contract DeusBridge is IDeusBridge, Ownable, Pausable {
         }
 
         if (fee[tokenId] > 0) {
-            uint feeAmount = amount * fee[tokenId] / scale;
+            uint256 feeAmount = (amount * fee[tokenId]) / scale;
             amount -= feeAmount;
             collectedFee[tokenId] += feeAmount;
         }
@@ -184,28 +197,42 @@ contract DeusBridge is IDeusBridge, Ownable, Pausable {
 
     function claim(
         address user,
-        uint amount,
-        uint fromChain,
-        uint toChain,
-        uint tokenId,
-        uint txId,
+        uint256 amount,
+        uint256 fromChain,
+        uint256 toChain,
+        uint256 tokenId,
+        uint256 txId,
         bytes calldata _reqId,
         SchnorrSign[] calldata sigs
     ) external {
-        require(sideContracts[fromChain] != address(0), 'Bridge: source contract not exist');
+        require(
+            sideContracts[fromChain] != address(0),
+            "Bridge: source contract not exist"
+        );
         require(toChain == network, "Bridge: toChain should equal network");
-        require(sigs.length >= minReqSigs, "Bridge: insufficient number of signatures");
+        require(
+            sigs.length >= minReqSigs,
+            "Bridge: insufficient number of signatures"
+        );
 
         {
             bytes32 hash = keccak256(
-            abi.encodePacked(
-                abi.encodePacked(sideContracts[fromChain], txId, tokenId, amount),
-                abi.encodePacked(fromChain, toChain, user, ETH_APP_ID)
+                abi.encodePacked(
+                    abi.encodePacked(
+                        sideContracts[fromChain],
+                        txId,
+                        tokenId,
+                        amount
+                    ),
+                    abi.encodePacked(fromChain, toChain, user, ETH_APP_ID)
                 )
             );
 
             IMuonV02 muon = IMuonV02(muonContract);
-            require(muon.verify(_reqId, uint(hash), sigs), "Bridge: not verified");
+            require(
+                muon.verify(_reqId, uint256(hash), sigs),
+                "Bridge: not verified"
+            );
         }
 
         require(!claimedTxs[fromChain][txId], "Bridge: already claimed");
@@ -217,7 +244,7 @@ contract DeusBridge is IDeusBridge, Ownable, Pausable {
             if (tokens[tokenId] == deiAddress) {
                 bridgeReserve += amount;
             }
-        } else { 
+        } else {
             token.transfer(user, amount);
         }
 
@@ -225,42 +252,52 @@ contract DeusBridge is IDeusBridge, Ownable, Pausable {
         emit Claim(user, tokenId, amount, fromChain, txId);
     }
 
-
     /* ========== VIEWS ========== */
 
     // This function use pool feature to handle buyback and recollateralize on DEI minter pool
-    function collatDollarBalance(uint collat_usd_price) public view returns (uint) {
-        uint collateralRatio = IDEIStablecoin(deiAddress).global_collateral_ratio();
-        return bridgeReserve * collateralRatio / 1e6;
+    function collatDollarBalance(uint256 collat_usd_price)
+        public
+        view
+        returns (uint256)
+    {
+        uint256 collateralRatio = IDEIStablecoin(deiAddress)
+            .global_collateral_ratio();
+        return (bridgeReserve * collateralRatio) / 1e6;
     }
 
-    function pendingTxs(
-        uint fromChain, 
-        uint[] calldata ids
-    ) public view returns (bool[] memory unclaimedIds) {
+    function pendingTxs(uint256 fromChain, uint256[] calldata ids)
+        public
+        view
+        returns (bool[] memory unclaimedIds)
+    {
         unclaimedIds = new bool[](ids.length);
-        for(uint i=0; i < ids.length; i++){
+        for (uint256 i = 0; i < ids.length; i++) {
             unclaimedIds[i] = claimedTxs[fromChain][ids[i]];
         }
     }
 
-    function getUserTxs(
-        address user, 
-        uint toChain
-    ) public view returns (uint[] memory) {
+    function getUserTxs(address user, uint256 toChain)
+        public
+        view
+        returns (uint256[] memory)
+    {
         return userTxs[user][toChain];
     }
 
-    function getTransaction(uint txId_) public view returns(
-        uint txId,
-        uint tokenId,
-        uint amount,
-        uint fromChain,
-        uint toChain,
-        address user,
-        uint txBlockNo,
-        uint currentBlockNo
-    ){
+    function getTransaction(uint256 txId_)
+        public
+        view
+        returns (
+            uint256 txId,
+            uint256 tokenId,
+            uint256 amount,
+            uint256 fromChain,
+            uint256 toChain,
+            address user,
+            uint256 txBlockNo,
+            uint256 currentBlockNo
+        )
+    {
         txId = txs[txId_].txId;
         tokenId = txs[txId_].tokenId;
         amount = txs[txId_].amount;
@@ -271,46 +308,51 @@ contract DeusBridge is IDeusBridge, Ownable, Pausable {
         currentBlockNo = block.number;
     }
 
-    function getExecutingChainID() public view returns (uint) {
-        uint id;
+    function getExecutingChainID() public view returns (uint256) {
+        uint256 id;
         assembly {
             id := chainid()
         }
         return id;
     }
 
-
     /* ========== RESTRICTED FUNCTIONS ========== */
 
-    function setBridgeReserve(uint bridgeReserve_) external onlyOwner {
+    function setBridgeReserve(uint256 bridgeReserve_) external onlyOwner {
         emit BridgeReserveSet(bridgeReserve, bridgeReserve_);
 
         bridgeReserve = bridgeReserve_;
     }
 
-    function setToken(uint tokenId, address tokenAddress) external onlyOwner {
+    function setToken(uint256 tokenId, address tokenAddress)
+        external
+        onlyOwner
+    {
         tokens[tokenId] = tokenAddress;
     }
 
-    function setNetworkID(uint network_) external onlyOwner {
+    function setNetworkID(uint256 network_) external onlyOwner {
         network = network_;
         delete sideContracts[network];
     }
 
-    function setFee(uint tokenId, uint fee_) external onlyOwner {
+    function setFee(uint256 tokenId, uint256 fee_) external onlyOwner {
         fee[tokenId] = fee_;
     }
-    
+
     function setDeiAddress(address deiAddress_) external onlyOwner {
         deiAddress = deiAddress_;
     }
 
-    function setMinReqSigs(uint minReqSigs_) external onlyOwner {
+    function setMinReqSigs(uint256 minReqSigs_) external onlyOwner {
         minReqSigs = minReqSigs_;
     }
 
-    function setSideContract(uint network_, address address_) external onlyOwner {
-        require (network != network_, "Bridge: current network");
+    function setSideContract(uint256 network_, address address_)
+        external
+        onlyOwner
+    {
+        require(network != network_, "Bridge: current network");
         sideContracts[network_] = address_;
     }
 
@@ -326,23 +368,34 @@ contract DeusBridge is IDeusBridge, Ownable, Pausable {
         muonContract = muonContract_;
     }
 
-    function pause() external onlyOwner { super._pause(); }
+    function pause() external onlyOwner {
+        super._pause();
+    }
 
-    function unpase() external onlyOwner { super._unpause(); }
+    function unpase() external onlyOwner {
+        super._unpause();
+    }
 
-    function withdrawFee(uint tokenId, address to) external onlyOwner {
+    function withdrawFee(uint256 tokenId, address to) external onlyOwner {
         require(collectedFee[tokenId] > 0, "Bridge: No fee to collect");
 
         IERC20(tokens[tokenId]).pool_mint(to, collectedFee[tokenId]);
         collectedFee[tokenId] = 0;
     }
 
-    function emergencyWithdrawETH(address to, uint amount) external onlyOwner {
+    function emergencyWithdrawETH(address to, uint256 amount)
+        external
+        onlyOwner
+    {
         require(to != address(0));
         payable(to).transfer(amount);
     }
 
-    function emergencyWithdrawERC20Tokens(address tokenAddr, address to, uint amount) external onlyOwner {
+    function emergencyWithdrawERC20Tokens(
+        address tokenAddr,
+        address to,
+        uint256 amount
+    ) external onlyOwner {
         require(to != address(0));
         IERC20(tokenAddr).transfer(to, amount);
     }
